route fix

+++ AppOneDrive/app/Http/Controllers/API/AuthController.php
@@ -10,11 +10,12 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
  
-
-
+ 
+ 
 class AuthController extends Controller
 {
-    public function register(Request $request){
+    public function register(Request $request)
+    {
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:255',
             'surname' => 'required|string|max:255',
@@ -22,8 +23,8 @@ class AuthController extends Controller
             'password' => 'required|string|min:8',
         ]);
  
-        if($validator->fails()){
-            return response()->json($validator->errors());
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
         }
  
         $user = User::create([
@@ -35,24 +36,35 @@ class AuthController extends Controller
  
         $token = $user->createToken('auth_token')->plainTextToken;
  
-        return response()->json(['message' => 'User created successfully.', 'user' => $user, 'auth_token' => $token, 'token_type' => 'Bearer']);
+        return response()->json(['message' => 'User created successfully.', 'userId' => $user->id, 'userName' => $user->name, 'auth_token' => $token, 'token_type' => 'Bearer'], 200);
     }
  
-    public function login(Request $request){
-        if(!Auth::attempt($request->only('email', 'password'))){
-            return response()->json(['message' => 'Unauthorized'], 401);
+    public function login(Request $request)
+    {
+        $validator = Validator::make($request->all(), [
+            'email' => 'required|string|email|max:255',
+            'password' => 'required|string|max:255'
+        ]);
+ 
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
+ 
+        if (!Auth::attempt($request->only('email', 'password'))) {
+            return response()->json(['message' => 'Invalid email or password.'], 401);
         }
  
         $user = User::where('email', $request['email'])->firstOrFail();
-        
+ 
         DB::table('personal_access_tokens')->where('tokenable_id', $user->id)->delete();
         $token = $user->createToken('auth_token')->plainTextToken;
  
-        return response()->json(['message' => 'Welcome ' . $user->name, 'auth_token' => $token, 'token_type' => 'Bearer']);
+        return response()->json(['message' => 'Welcome ' . $user->name, 'userId' => $user->id, 'userName' => $user->name, 'auth_token' => $token, 'token_type' => 'Bearer'], 200);
     }
-
-    public function logout(){
-        $user=User::find(auth()->id());
+ 
+    public function logout()
+    {
+        $user = User::find(auth()->id());
         DB::table('personal_access_tokens')->where('tokenable_id', auth()->id())->delete();
         return response()->json(['message' => 'Goodbye ' . $user->name]);
     }
